Alterando no da classe EventServiceImpl
